refactor: 리팩토링한 interceptor 교체
