+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -4,11 +4,11 @@ import nextstep.auth.authentication.BasicAuthenticationFilter;
 import nextstep.auth.authentication.BearerTokenAuthenticationFilter;
 import nextstep.auth.authentication.FormAuthenticationConverter;
 import nextstep.auth.authentication.TokenAuthenticationConverter;
-import nextstep.auth.authentication.UsernamePasswordAuthenticationFilter2;
+import nextstep.auth.authentication.UsernamePasswordAuthenticationFilter;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
 import nextstep.auth.context.SecurityContextPersistenceFilter;
 import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenAuthenticationInterceptor2;
+import nextstep.auth.token.TokenAuthenticationInterceptor;
 import nextstep.member.application.LoginMemberService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -29,11 +29,10 @@ public class AuthConfig implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(new SecurityContextPersistenceFilter());
-//        registry.addInterceptor(new UsernamePasswordAuthenticationFilter(loginMemberService)).addPathPatterns("/login/form");
-        registry.addInterceptor(new UsernamePasswordAuthenticationFilter2(loginMemberService, new FormAuthenticationConverter())).addPathPatterns(
-                "/login/form");
-//        registry.addInterceptor(new TokenAuthenticationInterceptor(loginMemberService, jwtTokenProvider)).addPathPatterns("/login/token");
-        registry.addInterceptor(new TokenAuthenticationInterceptor2(loginMemberService, jwtTokenProvider, new TokenAuthenticationConverter())).addPathPatterns("/login/token");
+        registry.addInterceptor(new UsernamePasswordAuthenticationFilter(loginMemberService,
+                new FormAuthenticationConverter())).addPathPatterns("/login/form");
+        registry.addInterceptor(new TokenAuthenticationInterceptor(loginMemberService, jwtTokenProvider,
+                new TokenAuthenticationConverter())).addPathPatterns("/login/token");
         registry.addInterceptor(new BasicAuthenticationFilter(loginMemberService));
         registry.addInterceptor(new BearerTokenAuthenticationFilter(jwtTokenProvider));
     }
