+++ src/main/java/nextstep/auth/authentication/UsernamePasswordAuthenticationFilter.java
@@ -4,49 +4,20 @@ import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContextHolder;
 import nextstep.member.application.LoginMemberService;
 import nextstep.member.domain.LoginMember;
-import org.springframework.web.servlet.HandlerInterceptor;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.Map;
 
-public class UsernamePasswordAuthenticationFilter implements HandlerInterceptor {
-    public static final String USERNAME_FIELD = "username";
-    public static final String PASSWORD_FIELD = "password";
+public class UsernamePasswordAuthenticationFilter extends NonChainAuthenticationInterceptor {
 
-    private LoginMemberService loginMemberService;
-
-    public UsernamePasswordAuthenticationFilter(LoginMemberService loginMemberService) {
-        this.loginMemberService = loginMemberService;
+    public UsernamePasswordAuthenticationFilter(LoginMemberService loginMemberService,
+                                                AuthenticationConverter authenticationGenerator) {
+        super(loginMemberService, authenticationGenerator);
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        try {
-            Map<String, String[]> paramMap = request.getParameterMap();
-            String username = paramMap.get(USERNAME_FIELD)[0];
-            String password = paramMap.get(PASSWORD_FIELD)[0];
-
-            AuthenticationToken token = new AuthenticationToken(username, password);
-
-            String principal = token.getPrincipal();
-            LoginMember loginMember = loginMemberService.loadUserByUsername(principal);
-
-            if (loginMember == null) {
-                throw new AuthenticationException();
-            }
-
-            if (!loginMember.checkPassword(token.getCredentials())) {
-                throw new AuthenticationException();
-            }
-
-            Authentication authentication = new Authentication(loginMember.getEmail(), loginMember.getAuthorities());
-
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-
-            return false;
-        } catch (Exception e) {
-            return true;
-        }
+    public void afterAuthentication(LoginMember loginMember, HttpServletResponse response) {
+        Authentication authentication = new Authentication(loginMember.getEmail(), loginMember.getAuthorities());
+        SecurityContextHolder.getContext().setAuthentication(authentication);
     }
+
 }
